+++ core-api/src/main/java/com/optimizely/ab/decision/audience/AudienceEvaluator.java
@@ -16,23 +16,22 @@
 package com.optimizely.ab.decision.audience;
 
 import com.optimizely.ab.config.Experiment;
+import com.optimizely.ab.config.ProjectConfig;
 import com.optimizely.ab.event.internal.UserContext;
-import com.optimizely.ab.internal.ExperimentUtils;
 
 import javax.annotation.Nonnull;
 
 /**
  * Determines whether a user satisfies audience conditions for the experiment.
  */
-public class AudienceEvaluator implements IAudienceEvaluator {
+public interface AudienceEvaluator {
     /**
      * Validate Audience conditions for the experiment
+     *
+     * @param experiment  The Experiment the user will be bucketed into.
+     * @param userContext It have user id, attributes and a reference to the current {@link ProjectConfig}
+     * @return flag if audience attributes are valid
      */
-    @Override
-    public boolean evaluate(@Nonnull Experiment experiment,
-                            @Nonnull UserContext userContext) {
-       return ExperimentUtils.isUserInExperiment(userContext.getProjectConfig(),
-                                                 experiment,
-                                                 userContext.getAttributes());
-    }
+    boolean evaluate(@Nonnull Experiment experiment,
+                     @Nonnull UserContext userContext);
 }
